Fix integer array indices

+++ SLIT/Lens.py
@@ -123,7 +123,7 @@ def alpha_def(kappa, n1,n2,extra=0):
     intex = scp.fftconvolve(tabx, (kappa), mode = 'same')/np.pi
     intey = scp.fftconvolve(taby, (kappa), mode = 'same')/np.pi
 
-    return intex[x0-(n1)/2:x0+(n1)/2,y0-(n2)/2:y0+(n2)/2], intey[x0-(n1)/2:x0+(n1)/2,y0-(n2)/2:y0+(n2)/2]
+    return intex[int(x0-(n1)/2):int(x0+(n1)/2),int(y0-(n2)/2):int(y0+(n2)/2)], intey[int(x0-(n1)/2):int(x0+(n1)/2),int(y0-(n2)/2):int(y0+(n2)/2)]
 
 
 def beta(kappa,theta):
